+++ zetaclient/chains/base/observer_test.go
@@ -1,4 +1,4 @@
-package base_test
+package base
 
 import (
 	"fmt"
@@ -11,12 +11,12 @@ import (
 	"github.com/zeta-chain/node/pkg/chains"
 	"github.com/zeta-chain/node/testutil/sample"
 	observertypes "github.com/zeta-chain/node/x/observer/types"
-	"github.com/zeta-chain/node/zetaclient/chains/base"
-	"github.com/zeta-chain/node/zetaclient/chains/interfaces"
+	"github.com/zeta-chain/node/zetaclient/chains/tssrepo"
 	"github.com/zeta-chain/node/zetaclient/chains/zrepo"
 	"github.com/zeta-chain/node/zetaclient/config"
 	zctx "github.com/zeta-chain/node/zetaclient/context"
 	"github.com/zeta-chain/node/zetaclient/db"
+	"github.com/zeta-chain/node/zetaclient/keys"
 	"github.com/zeta-chain/node/zetaclient/mode"
 	"github.com/zeta-chain/node/zetaclient/testutils/mocks"
 )
@@ -27,7 +27,7 @@ const (
 )
 
 type testSuite struct {
-	*base.Observer
+	*Observer
 	db       *db.DB
 	tss      *mocks.TSS
 	zetacore *mocks.ZetacoreClient
@@ -64,20 +64,20 @@ func newTestSuite(t *testing.T, chain chains.Chain, opts ...opt) *testSuite {
 	if testOpts.ConfirmationParams != nil {
 		chainParams.ConfirmationParams = testOpts.ConfirmationParams
 	}
-	zetacoreClient := mocks.NewZetacoreClient(t)
+	zetacoreClient := mocks.NewZetacoreClient(t).WithKeys(&keys.Keys{}).WithZetaChain()
 	tss := mocks.NewTSS(t)
 
 	database := createDatabase(t)
 
 	// create observer
 	zetaRepo := zrepo.New(zetacoreClient, chain, mode.StandardMode)
-	logger := base.DefaultLogger()
-	ob, err := base.NewObserver(
+	logger := DefaultLogger()
+	ob, err := NewObserver(
 		chain,
 		chainParams,
 		zetaRepo,
 		tss,
-		base.DefaultBlockCacheSize,
+		DefaultBlockCacheSize,
 		nil,
 		database,
 		logger,
@@ -99,7 +99,7 @@ func TestNewObserver(t *testing.T) {
 	appContext := zctx.New(config.New(false), nil, zerolog.Nop())
 	zetacoreClient := mocks.NewZetacoreClient(t)
 	tss := mocks.NewTSS(t)
-	blockCacheSize := base.DefaultBlockCacheSize
+	blockCacheSize := DefaultBlockCacheSize
 
 	database := createDatabase(t)
 
@@ -110,7 +110,7 @@ func TestNewObserver(t *testing.T) {
 		chainParams    observertypes.ChainParams
 		appContext     *zctx.AppContext
 		zetacoreClient zrepo.ZetacoreClient
-		tssSigner      interfaces.TSSSigner
+		tssSigner      tssrepo.TSSClient
 		blockCacheSize int
 		fail           bool
 		message        string
@@ -141,7 +141,7 @@ func TestNewObserver(t *testing.T) {
 	// run tests
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			ob, err := base.NewObserver(
+			ob, err := NewObserver(
 				tt.chain,
 				tt.chainParams,
 				zrepo.New(tt.zetacoreClient, tt.chain, mode.StandardMode),
@@ -149,7 +149,7 @@ func TestNewObserver(t *testing.T) {
 				tt.blockCacheSize,
 				nil,
 				database,
-				base.DefaultLogger(),
+				DefaultLogger(),
 			)
 			if tt.fail {
 				require.ErrorContains(t, err, tt.message)
@@ -305,7 +305,7 @@ func TestOutboundID(t *testing.T) {
 
 func TestLoadLastBlockScanned(t *testing.T) {
 	chain := chains.Ethereum
-	envvar := base.EnvVarLatestBlockByChain(chain)
+	envvar := EnvVarLatestBlockByChain(chain)
 
 	t.Run("should be able to load last block scanned", func(t *testing.T) {
 		// create observer and open db
@@ -355,7 +355,7 @@ func TestLoadLastBlockScanned(t *testing.T) {
 		ob.WriteLastBlockScannedToDB(100)
 
 		// set env var to 'latest'
-		os.Setenv(envvar, base.EnvVarLatestBlock)
+		os.Setenv(envvar, EnvVarLatestBlock)
 
 		// last block scanned should remain 0
 		err := ob.LoadLastBlockScanned()
@@ -421,7 +421,7 @@ func TestReadWriteDBLastBlockScanned(t *testing.T) {
 }
 func TestLoadLastTxScanned(t *testing.T) {
 	chain := chains.SolanaDevnet
-	envvar := base.EnvVarLatestTxByChain(chain)
+	envvar := EnvVarLatestTxByChain(chain)
 	lastTx := "5LuQMorgd11p8GWEw6pmyHCDtA26NUyeNFhLWPNk2oBoM9pkag1LzhwGSRos3j4TJLhKjswFhZkGtvSGdLDkmqsk"
 
 	t.Run("should be able to load last tx scanned", func(t *testing.T) {