+++ zetaclient/chains/bitcoin/observer/observer_test.go
@@ -23,7 +23,7 @@ import (
 	"github.com/zeta-chain/node/x/crosschain/types"
 	observertypes "github.com/zeta-chain/node/x/observer/types"
 	"github.com/zeta-chain/node/zetaclient/chains/base"
-	"github.com/zeta-chain/node/zetaclient/chains/interfaces"
+	"github.com/zeta-chain/node/zetaclient/chains/tssrepo"
 	"github.com/zeta-chain/node/zetaclient/chains/zrepo"
 	"github.com/zeta-chain/node/zetaclient/metrics"
 	"github.com/zeta-chain/node/zetaclient/testutils/mocks"
@@ -84,7 +84,7 @@ func Test_NewObserver(t *testing.T) {
 		btcClient      *mocks.BitcoinClient
 		chainParams    observertypes.ChainParams
 		zetacoreClient zrepo.ZetacoreClient
-		tssSigner      interfaces.TSSSigner
+		tssSigner      tssrepo.TSSClient
 		logger         base.Logger
 		ts             *metrics.TelemetryServer
 		errorMessage   string
@@ -299,7 +299,7 @@ func newTestSuite(t *testing.T, chain chains.Chain, opts ...opt) *testSuite {
 	client := mocks.NewBitcoinClient(t)
 	zetacore := mocks.NewZetacoreClient(t)
 
-	var tssSigner interfaces.TSSSigner
+	var tssSigner tssrepo.TSSClient
 	if chains.IsBitcoinMainnet(chain.ChainId) {
 		tssSigner = mocks.NewTSS(t).FakePubKey(testutils.TSSPubKeyMainnet)
 	} else {