+++ zetaclient/chains/evm/evm.go
@@ -88,6 +88,7 @@ func (e *EVM) Start(ctx context.Context) error {
 	// Observers
 	register(e.observer.ObserveInbound, "observe_inbound", optInboundInterval, optInboundSkipper)
 	register(e.observer.ProcessInboundTrackers, "process_inbound_trackers", optInboundInterval, optInboundSkipper)
+	register(e.observer.ProcessInternalTrackers, "process_internal_trackers", optInboundInterval, optInboundSkipper)
 	register(e.observer.PostGasPrice, "post_gas_price", optGasInterval, optGasPriceSkipper)
 	register(e.observer.CheckRPCStatus, "check_rpc_status")
 	register(e.observer.ProcessOutboundTrackers, "process_outbound_trackers", optOutboundInterval, optOutboundSkipper)