+++ zetaclient/chains/sui/sui.go
@@ -89,6 +89,7 @@ func (s *Sui) Start(ctx context.Context) error {
 	register(s.observer.ObserveGasPrice, "observe_gas_price", optGasInterval, optGasPriceSkipper)
 	register(s.observer.ObserveInbound, "observe_inbounds", optInboundInterval, optInboundSkipper)
 	register(s.observer.ProcessInboundTrackers, "process_inbound_trackers", optInboundInterval, optInboundSkipper)
+	register(s.observer.ProcessInternalTrackers, "process_internal_trackers", optInboundInterval, optInboundSkipper)
 	register(s.observer.ProcessOutboundTrackers, "process_outbound_trackers", optOutboundInterval, optOutboundSkipper)
 
 	// CCTX scheduler (every zetachain block)
@@ -143,9 +144,8 @@ func (s *Sui) scheduleCCTX(ctx context.Context) error {
 		maxNonce   = firstNonce + lookback
 	)
 
-	for i := range cctxList {
+	for i, cctx := range cctxList {
 		var (
-			cctx           = cctxList[i]
 			outboundID     = base.OutboundIDFromCCTX(cctx)
 			outboundParams = cctx.GetCurrentOutboundParam()
 			nonce          = outboundParams.TssNonce
@@ -182,7 +182,6 @@ func (s *Sui) scheduleCCTX(ctx context.Context) error {
 			if err := s.signer.ProcessCCTX(ctx, cctx, zetaHeight); err != nil {
 				s.outboundLogger(outboundID).Error().Err(err).Msg("error calling ProcessCCTX")
 			}
-
 			return nil
 		})
 	}