+++ zetaclient/zetacore/client_monitor.go
@@ -9,6 +9,7 @@ import (
 	"github.com/pkg/errors"
 
 	"github.com/zeta-chain/node/pkg/constant"
+	zetaerrors "github.com/zeta-chain/node/pkg/errors"
 	"github.com/zeta-chain/node/pkg/retry"
 	"github.com/zeta-chain/node/x/crosschain/types"
 	"github.com/zeta-chain/node/zetaclient/logs"
@@ -102,6 +103,7 @@ func (c *Client) MonitorVoteInboundResult(
 	zetaTxHash string,
 	retryGasLimit uint64,
 	msg *types.MsgVoteInbound,
+	monitorErrCh chan<- zetaerrors.ErrTxMonitor,
 ) error {
 	logger := c.logger.With().Str(logs.FieldZetaTx, zetaTxHash).Logger()
 
@@ -112,15 +114,19 @@ func (c *Client) MonitorVoteInboundResult(
 	}()
 
 	call := func() error {
-		err := c.monitorVoteInboundResult(ctx, zetaTxHash, retryGasLimit, msg)
+		err := c.monitorVoteInboundResult(ctx, zetaTxHash, retryGasLimit, msg, monitorErrCh)
 
 		// force retry on err
 		return retry.Retry(err)
 	}
 
+	//                               10 attempts,    2 seconds,     4 seconds max
+	// This will retry for a maximum of ~40 seconds with exponential backoff,
+	// However, this call is recursive for up to 1 layer and so the maxim time this can take is ~80 seconds
 	err := retryWithBackoff(call, monitorRetryCount, monitorInterval, monitorInterval*2)
 	if err != nil {
-		logger.Error().Err(err).Msg("unable to query tx result")
+		// All errors are forced to be retryable, we only return an error if the tx result cannot be queried
+		logger.Error().Err(err).Str(logs.FieldZetaTx, zetaTxHash).Msg("unable to query tx result")
 		return err
 	}
 
@@ -132,6 +138,7 @@ func (c *Client) monitorVoteInboundResult(
 	zetaTxHash string,
 	retryGasLimit uint64,
 	msg *types.MsgVoteInbound,
+	monitorErrCh chan<- zetaerrors.ErrTxMonitor,
 ) error {
 	// query tx result from ZetaChain
 	txResult, err := c.QueryTxResult(zetaTxHash)
@@ -141,6 +148,8 @@ func (c *Client) monitorVoteInboundResult(
 
 	logger := c.logger.With().Str("inbound_raw_log", txResult.RawLog).Logger()
 
+	// There is no error returned from here which mean the MonitorVoteInboundResult would return nil and no error is posted to monitorErrCh
+	// However the channel is passed to the subsequent call, which can post an error to the channel if the "execute" vote fails.
 	switch {
 	case strings.Contains(txResult.RawLog, "failed to execute message"):
 		// the inbound vote tx shouldn't fail to execute. this shouldn't happen
@@ -155,7 +164,7 @@ func (c *Client) monitorVoteInboundResult(
 		logger.Debug().Str(logs.FieldZetaTx, zetaTxHash).Msg("out of gas")
 		if retryGasLimit > 0 {
 			// new retryGasLimit set to 0 to prevent reentering this function
-			resentZetaTxHash, _, err := c.PostVoteInbound(ctx, retryGasLimit, 0, msg)
+			resentZetaTxHash, _, err := c.PostVoteInbound(ctx, retryGasLimit, 0, msg, monitorErrCh)
 			if err != nil {
 				logger.Error().Err(err).Str(logs.FieldZetaTx, zetaTxHash).Msg("failed to resend tx")
 			} else {