+++ changelog.md
@@ -35,6 +35,11 @@ Also EnableSolanaAddressLookupTable feature flag should be set.
 * [4274](https://github.com/zeta-chain/node/pull/4274) - multiple evm calls in single tx
 * [4266](https://github.com/zeta-chain/node/pull/4266) - add support for Solana Address Lookup Table in withdraw and call
 * [4288](https://github.com/zeta-chain/node/pull/4288) - zetaclient config feature flag for multiple evm calls
+* [4313](https://github.com/zeta-chain/node/pull/4313) - add dry-mode to Sui and add client wrappers to Sui and TON
+* [4317](https://github.com/zeta-chain/node/pull/4317) - add dry-mode to Solana
+* [4325](https://github.com/zeta-chain/node/pull/4325) - add dry-mode to Bitcoin
+* [4326](https://github.com/zeta-chain/node/pull/4326) - add dry-mode to EVM
+* [4330](https://github.com/zeta-chain/node/pull/4330) - add TSS client dry-wrapper
 
 ### Tests
 
@@ -53,6 +58,7 @@ Also EnableSolanaAddressLookupTable feature flag should be set.
 
 * [4153](https://github.com/zeta-chain/node/pull/4153) - make the gas limit used for gateway calls a configurable parameter
 * [4277](https://github.com/zeta-chain/node/pull/4277) - add dry-mode support for TON
+* [4295](https://github.com/zeta-chain/node/pull/4295) - monitor failed inbound votes and add them to an internal inbound tracker cache
 
 ### Refactor